goals

+++ app/Http/Controllers/GiftController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Gift;
 use Illuminate\Http\Request;
 use Log;
+use App\Models\Goal;
 
 class GiftController extends Controller
 {
@@ -24,24 +25,29 @@ class GiftController extends Controller
         $variant['price'] = 0;
         $gift['variants'][] = $variant;
         $new_product = $shopifyApi->Product->post($gift);
-
-        Gift::create([
-            'shop_id' => $shop->id,
-            'product_id'=> $new_product['id'],
-            'product_title'=> $request->title,
-            'variant_id' => $variant['id'],
-            'variant_title' => $variant['title'],
-            'product_handle'=> $new_product['handle'],
-            'product_image'=>$request->image,
-        ]);
-        return response()-> json([
-            'status' => true
-        ]);
+        try {
+            Gift::create([
+                'shop_id' => $shop->id,
+                'product_id'=> $new_product['id'],
+                'product_title'=> $request->title,
+                'variant_id' => $variant['id'],
+                'variant_title' => $variant['title'],
+                'product_handle'=> $new_product['handle'],
+                'product_image'=>$request->image,
+            ]);
+            return response()-> json([
+                'status' => true
+            ]);
+        }catch(Exception $e) {
+            return response()-> json([
+                'status' => false
+            ]);
+        }
     }
 
     public function gifts(Request $request){
         $shop = $request->get('shop');
-        $gifts = Gift::where('shop_id',$shop->id)->get();
+        $gifts = $shop->gifts;
         return response()-> json($gifts);
     }
    
@@ -49,11 +55,24 @@ class GiftController extends Controller
     {
         $shop = $request->get('shop');
         $gift = Gift::where('product_id',$request->product_id)->first();
-        $gift->delete();
-        $shopifyApi = $shop->getShopifyApi();
-        $product = $shopifyApi->Product($request->product_id)->delete();
-        return response()-> json([
-            'status' => true
-        ]);
+        $goal = Goal::where('gift_id',$gift->id)->first();
+        if(!empty($goal)){
+            return response()-> json([
+                'status' => false,
+                'message' => 'The gift is being used'
+            ]);
+        }
+        try {
+            $gift->deleteData();
+            $shopifyApi = $shop->getShopifyApi();
+            $product = $shopifyApi->Product($request->product_id)->delete();
+            return response()-> json([
+                'status' => true
+            ]);
+        }catch(Exception $e) {
+            return response()-> json([
+                'status' => false
+            ]);
+        }
     }
 }
